requirements.txt
flask==3.0.3
matplotlib==3.9.0
pandas==2.2.2
scikit-learn==1.5.0
sqlalchemy==2.0.30
gunicorn
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai==0.8.3